requirements.txt
asgiref==3.4.1
async-timeout==4.0.2
autopep8==1.5.7
beautifulsoup4==4.10.0
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.6
cloudinary==1.26.0
crispy-bootstrap5==0.5
cryptography==35.0.0
cssselect2==0.5.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2
django-allauth==0.45.0
django-ckeditor==6.1.0
django-cloudinary-storage==0.3.0
django-crispy-forms==1.13.0
django-filer==2.1.2
django-import-export==2.8.0
django-js-asset==1.2.2
django-mptt==0.13.4
django-multiurl==1.5.0
django-polymorphic==3.0.0
django-ranged-response==0.2.0
django-redis==5.2.0
django-rename-app==0.1.3
django-simple-captcha==0.5.14
django-tinymce==3.7.1
easy-thumbnails==2.8.1
et-xmlfile==1.1.0
filelock==3.6.0
flake8==3.9.2
gunicorn==20.1.0
idna==3.2
lxml==4.9.2
MarkupPy==1.14
mccabe==0.6.1
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
Pillow==9.4.0
platformdirs==2.5.1
psycopg2==2.9.5
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.3.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0
redis==4.3.4
reportlab==3.6.12
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
svglib==1.2.0
tablib==3.2.0
tinycss2==1.1.1
toml==0.10.2
tzdata==2022.2
Unidecode==1.1.2
urllib3==1.26.6
virtualenv==20.13.1
webencodings==0.5.1
whitenoise==5.3.0
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
